Move vendor-specific client examples into subpackages
Vendor-targeted examples (Prosys, IJT, Unified Automation) now live in
their own subpackages instead of the flat client package. KeyStoreLoader
is made public so the subpackaged examples can access it.
