+++ milo-examples/client-examples/src/main/java/org/eclipse/milo/examples/client/KeyStoreLoader.java
@@ -28,7 +28,7 @@ import org.eclipse.milo.opcua.stack.core.util.SelfSignedCertificateGenerator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-class KeyStoreLoader {
+public class KeyStoreLoader {
 
   private static final Pattern IP_ADDR_PATTERN =
       Pattern.compile("^(([01]?\\d\\d?|2[0-4]\\d|25[0-5])\\.){3}([01]?\\d\\d?|2[0-4]\\d|25[0-5])$");
