+++ milo-examples/client-examples/src/main/java/org/eclipse/milo/examples/client/prosys/ProsysHistoryReadExample.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 the Eclipse Milo Authors
+ * Copyright (c) 2026 the Eclipse Milo Authors
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -8,13 +8,15 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.milo.examples.client;
+package org.eclipse.milo.examples.client.prosys;
 
 import static org.eclipse.milo.opcua.stack.core.types.builtin.unsigned.Unsigned.uint;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
+import org.eclipse.milo.examples.client.ClientExample;
+import org.eclipse.milo.examples.client.ClientExampleRunner;
 import org.eclipse.milo.opcua.sdk.client.OpcUaClient;
 import org.eclipse.milo.opcua.stack.core.types.builtin.ByteString;
 import org.eclipse.milo.opcua.stack.core.types.builtin.DataValue;
@@ -31,10 +33,10 @@ import org.eclipse.milo.opcua.stack.core.types.structured.HistoryReadValueId;
 import org.eclipse.milo.opcua.stack.core.types.structured.ReadRawModifiedDetails;
 import org.eclipse.milo.opcua.stack.core.util.Lists;
 
-public class HistoryReadExampleProsys implements ClientExample {
+public class ProsysHistoryReadExample implements ClientExample {
 
   public static void main(String[] args) throws Exception {
-    HistoryReadExampleProsys example = new HistoryReadExampleProsys();
+    ProsysHistoryReadExample example = new ProsysHistoryReadExample();
 
     new ClientExampleRunner(example, false).run();
   }
